requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
Flask-Login>=0.6.0
Flask-Migrate>=4.0
Jinja2>=3.0
reportlab>=4.0
Pillow>=9.0
click>=8.0
python-dotenv>=1.0
pytest>=7.0
requests>=2.28
stripe>=7.0
psycopg2-binary>=2.9
Werkzeug>=2.0
passlib>=1.7.4
email-validator>=1.3
gunicorn>=21.0
